Update proxy tests to use air-gapped nodes
