+++ framework/set/resources/proxy/rancher/upgrade.sh
@@ -7,7 +7,7 @@ RANCHER_TAG_VERSION=$4
 RANCHER_IMAGE=$5
 BASTION=$6
 RANCHER_AGENT_IMAGE=${7}
-PROXY_PORT="3128"
+PROXY_PORT="3228"
 NO_PROXY="localhost\\,127.0.0.0/8\\,10.0.0.0/8\\,172.0.0.0/8\\,192.168.0.0/16\\,.svc\\,.cluster.local\\,cattle-system.svc\\,169.254.169.254"
 
 set -ex
