+++ framework/set/resources/proxy/rke2/init-server.sh
@@ -3,14 +3,38 @@
 USER=$1
 GROUP=$2
 K8S_VERSION=$3
-RKE2_SERVER_IP=$4
+RKE2_SERVER_ONE_IP=$4
 RKE2_TOKEN=$5
 BASTION=$6
-PORT="3128"
+PORT="3228"
+PEM_FILE=/home/$USER/keyfile.pem
 
 set -e
 
-cat <<EOF | sudo tee /etc/default/rke2-server > /dev/null
+runSSH() {
+  local server="$1"
+  local cmd="$2"
+  
+  ssh -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null -i "$PEM_FILE" "$USER@$server" \
+  "export USER=${USER}; \
+   export GROUP=${GROUP}; \
+   export RKE2_SERVER_ONE_IP=${RKE2_SERVER_ONE_IP}; \
+   export RKE2_TOKEN=${RKE2_TOKEN}; \
+   export BASTION=${BASTION} \
+   export PORT=${PORT}; ${cmd}"
+}
+
+setupConfig() {
+    sudo mkdir -p /etc/rancher/rke2
+    sudo tee /etc/rancher/rke2/config.yaml > /dev/null << EOF
+token: ${RKE2_TOKEN}
+tls-san:
+  - ${RKE2_SERVER_ONE_IP}
+EOF
+}
+
+setupProxy() {
+  cat <<EOF | sudo tee /etc/default/rke2-server > /dev/null
 HTTP_PROXY=http://${BASTION}:${PORT}
 HTTPS_PROXY=http://${BASTION}:${PORT}
 NO_PROXY=localhost,127.0.0.0/8,10.0.0/8,cattle-system.svc,172.16.0.0/12,192.168.0.0/16,.svc,.cluster.local
@@ -20,21 +44,25 @@ CONTAINERD_NO_PROXY=localhost,127.0.0.0/8,10.0.0.0/8,172.16.0.0/12,192.168.0.0/1
 http_proxy=http://${BASTION}:${PORT}
 https_proxy=http://${BASTION}:${PORT}
 EOF
+}
 
-sudo mkdir -p /etc/rancher/rke2
-sudo touch /etc/rancher/rke2/config.yaml
+runSSH "${RKE2_SERVER_ONE_IP}" "sudo mv /home/${USER}/kubectl /usr/local/bin/"
 
-echo "token: ${RKE2_TOKEN}
-tls-san:
-  - ${RKE2_SERVER_IP}" | sudo tee /etc/rancher/rke2/config.yaml > /dev/null
+configFunction=$(declare -f setupConfig)
+runSSH "${RKE2_SERVER_ONE_IP}" "${configFunction}; setupConfig"
 
-curl -sfL https://get.rke2.io --output install.sh
-sudo chmod +x install.sh
-sudo INSTALL_RKE2_VERSION=${K8S_VERSION} INSTALL_RKE2_TYPE='server' ./install.sh
+setupProxyFunction=$(declare -f setupProxy)
+runSSH "${RKE2_SERVER_ONE_IP}" "${setupProxyFunction}; setupProxy"
 
-sudo systemctl enable rke2-server
-sudo systemctl start rke2-server
+runSSH "${RKE2_SERVER_ONE_IP}" "sudo INSTALL_RKE2_ARTIFACT_PATH=/home/${USER} sh install.sh"
+runSSH "${RKE2_SERVER_ONE_IP}" "sudo systemctl enable rke2-server"
+runSSH "${RKE2_SERVER_ONE_IP}" "sudo systemctl start rke2-server"
 
-sudo mkdir -p /home/${USER}/.kube
-sudo cp /etc/rancher/rke2/rke2.yaml /home/${USER}/.kube/config
-sudo chown -R ${USER}:${GROUP} /home/${USER}/.kube
+runSSH "${RKE2_SERVER_ONE_IP}" "sudo mkdir -p /home/${USER}/.kube"
+runSSH "${RKE2_SERVER_ONE_IP}" "sudo cp /etc/rancher/rke2/rke2.yaml /home/${USER}/.kube/config"
+runSSH "${RKE2_SERVER_ONE_IP}" "sudo chown -R ${USER}:${GROUP} /home/${USER}/.kube"
+
+mkdir -p ~/.kube
+ssh -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null -i ${PEM_FILE} ${USER}@${RKE2_SERVER_ONE_IP} "sudo cat /home/${USER}/.kube/config" > ~/.kube/config
+sed -i "s|server: https://127.0.0.1:6443|server: https://${RKE2_SERVER_ONE_IP}:6443|" ~/.kube/config
+kubectl get nodes
