+++ framework/set/resources/proxy/squid/squid.conf
@@ -27,17 +27,16 @@ http_access deny !Safe_ports
 
 http_access deny CONNECT !SSL_ports
 
-http_access allow localhost manager
 http_access deny manager
 
 http_access allow localnet
 http_access allow localhost
-
 http_access allow all
 
-http_port 3128
+http_port 3228
 
 coredump_dir /var/cache/squid
+cache_dir ufs /var/cache/squid 10000 16 256
 
 refresh_pattern ^ftp:		1440	20%	10080
 refresh_pattern ^gopher:	1440	0%	1440
