+++ tests/proxy/README.md
@@ -74,6 +74,8 @@ terraform:
   standalone:
     bootstrapPassword: ""                         # REQUIRED - this is the same as the adminPassword above, make sure they match
     certManagerVersion: ""                        # REQUIRED - (e.g. v1.15.3)
+    osGroup: ""                                   # REQUIRED - fill with desired value
+    osUser: ""                                    # REQUIRED - fill with desired value
     rancherAgentImage: ""                         # OPTIONAL - fill out only if you are using Rancher Prime or staging registry
     rancherChartVersion: ""                       # REQUIRED - fill with desired value
     rancherChartRepository: ""                    # REQUIRED - fill with desired value. Must end with a trailing /
@@ -83,10 +85,17 @@ terraform:
     repo: ""                                      # REQUIRED - fill with desired value
     rke2Group: ""                                 # REQUIRED - fill with group of the instance created
     rke2User: ""                                  # REQUIRED - fill with username of the instance created
-    rke2Version: ""                               # REQUIRED - fill with desired RKE2 k8s value (i.e. v1.30.6+rke2r1)
+    rke2Version: ""                               # REQUIRED - fill with desired RKE2 k8s value (i.e. v1.30.9)
     upgradedRancherChartRepository: ""            # OPTIONAL - set if upgraded. Fill with desired value. Must end with a trailing /
     upgradedRancherImage: ""                      # OPTIONAL - set if upgraded. Fill with desired value
     upgradedRancherTagVersion: ""                 # OPTIONAL - set if upgraded. Fill with desired value
+  ####################################
+  # STANDALONE CONFIG - REGISTRY SETUP
+  ####################################
+  standaloneRegistry:
+    registryName: ""                              # REQUIRED - fill with desired value
+    registryPassword: ""                          # REQUIRED - fill with desired value
+    registryUsername: ""                          # REQUIRED - fill with desired value
 #######################
 # TERRATEST CONFIG
 #######################
